+++ Niffler/NifflerUITests/Pages/BasePage.swift
@@ -1,0 +1,9 @@
+import XCTest
+
+class BasePage {
+    init(app: XCUIApplication) {
+        self.app = app
+    }
+    
+    let app: XCUIApplication
+}
