+++ Niffler/NifflerUITests/Pages/LoginPage.swift
@@ -1,0 +1,41 @@
+import XCTest
+
+class LoginPage: BasePage {
+    
+    @discardableResult
+    func input(login: String, password: String) -> Self {
+        input(login: login)
+        input(password: password)
+        pressLoginButton()
+        return self
+    }
+    
+    private func input(login: String) {
+        XCTContext.runActivity(named: "Ввожу логин \(login)") { _ in
+            app.textFields["userNameTextField"].tap()
+            app.textFields["userNameTextField"].typeText(login)
+        }
+    }
+    
+    private func input(password: String) {
+        XCTContext.runActivity(named: "Ввожу пароль \(password)") { _ in
+            app.secureTextFields["passwordTextField"].tap()
+            app.secureTextFields["passwordTextField"].typeText(password)
+        }
+    }
+    
+    private func pressLoginButton() {
+        XCTContext.runActivity(named: "Жму кнопку логина") { _ in
+            app.buttons["loginButton"].tap()
+        }
+    }
+    
+    func assertIsLoginErrorShown(file: StaticString = #filePath, line: UInt = #line) {
+        XCTContext.runActivity(named: "Жду сообщение с ошибкой") { _ in
+            let isFound = app.staticTexts["Нет такого пользователя. Попробуйте другие данные"].waitForExistence(timeout: 5)
+            XCTAssertTrue(isFound,
+                          "Не нашли сообщение о неправильном логине",
+                          file: file, line: line)
+        }
+    }
+}
